+++ cookietemple/custom_cookietemple_cli/custom_click.py
@@ -1,10 +1,10 @@
 import click
 
-from cookietemple.cli_tools.suggest_similar_commands import MAIN_COMMANDS
+from cookietemple.custom_cookietemple_cli.suggest_similar_commands import MAIN_COMMANDS
 from cookietemple.info.levensthein_dist import most_similar_command
 
 
-class CustomHelpAndCommands(click.Group):
+class HelpErrorHandling(click.Group):
     """
     Customise the order of subcommands for --help
     https://stackoverflow.com/a/47984810/713980
@@ -12,15 +12,15 @@ class CustomHelpAndCommands(click.Group):
 
     def __init__(self, *args, **kwargs):
         self.help_priorities = {}
-        super(CustomHelpAndCommands, self).__init__(*args, **kwargs)
+        super(HelpErrorHandling, self).__init__(*args, **kwargs)
 
     def get_help(self, ctx):
         self.list_commands = self.list_commands_for_help
-        return super(CustomHelpAndCommands, self).get_help(ctx)
+        return super(HelpErrorHandling, self).get_help(ctx)
 
     def list_commands_for_help(self, ctx):
         """reorder the list of commands when listing the help"""
-        commands = super(CustomHelpAndCommands, self).list_commands(ctx)
+        commands = super(HelpErrorHandling, self).list_commands(ctx)
         return (c[1] for c in sorted((self.help_priorities.get(command, 1000), command) for command in commands))
 
     def command(self, *args, **kwargs):
@@ -31,7 +31,7 @@ class CustomHelpAndCommands(click.Group):
         help_priorities = self.help_priorities
 
         def decorator(f):
-            cmd = super(CustomHelpAndCommands, self).command(*args, **kwargs)(f)
+            cmd = super(HelpErrorHandling, self).command(*args, **kwargs)(f)
             help_priorities[cmd.name] = help_priority
             return cmd
