+++ cookietemple/info/info.py
@@ -7,7 +7,7 @@ from tabulate import tabulate
 from cookietemple.info.levensthein_dist import most_similar_command
 from cookietemple.list.list import load_available_templates
 from cookietemple.util.dict_util import is_nested_dictionary
-from cookietemple.cli_tools.suggest_similar_commands import load_available_handles
+from cookietemple.custom_cookietemple_cli.suggest_similar_commands import load_available_handles
 
 WD = os.path.dirname(__file__)
 TEMPLATES_PATH = f'{WD}/../create/templates'
