+++ cookietemple/info/levensthein_dist.py
@@ -1,4 +1,4 @@
-from cookietemple.cli_tools.suggest_similar_commands import SIMILARITY_FACTOR
+from cookietemple.custom_cookietemple_cli.suggest_similar_commands import SIMILARITY_FACTOR
 
 
 def levensthein_dist(input_command: str, candidate: str) -> int:
